# Repository_health_analyser/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RepoHealth API — git ingestion, code metrics, knowledge graphs, and LLM insights
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repos:
    get:
      operationId: listRepos
      tags: [repos]
      summary: List all ingested repositories
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repo"
    post:
      operationId: ingestRepo
      tags: [repos]
      summary: Ingest a new Git repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoInput"
      responses:
        "202":
          description: Ingestion started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repo"
        "400":
          description: Invalid repository URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /repos/{repoId}/stats:
    get:
      operationId: getRepoStats
      tags: [repos]
      summary: Get aggregate stats for the repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Aggregate repository statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoStats"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/commits:
    get:
      operationId: listCommits
      tags: [metrics]
      summary: List all commits with their health metrics
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of commits with metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitPage"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/commits/{sha}:
    get:
      operationId: getCommit
      tags: [metrics]
      summary: Get a single commit with detailed metrics
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: sha
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Commit with detailed metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitDetail"
        "404":
          description: Commit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/timeline:
    get:
      operationId: getTimeline
      tags: [metrics]
      summary: Get health metric timeline across all commits
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Time-series health metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelinePoint"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/hotspots:
    get:
      operationId: getHotspots
      tags: [metrics]
      summary: Get hotspot risk map — files ranked by churn × coupling
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File hotspot risk scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HotspotFile"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/riskiest-files:
    get:
      operationId: getRiskiestFiles
      tags: [metrics]
      summary: Get top N riskiest files with composite risk scores
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Top risky files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HotspotFile"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RepoInput:
      type: object
      properties:
        url:
          type: string
          description: Git repository URL (GitHub HTTPS URL)
        name:
          type: string
          description: Optional display name for the repo
        maxCommits:
          type: ["integer", "null"]
          description: Maximum commits to process (default 500)
      required:
        - url
    Repo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [pending, ingesting, ready, error]
        totalCommits:
          type: ["integer", "null"]
        processedCommits:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        errorMessage:
          type: ["string", "null"]
      required:
        - id
        - name
        - url
        - status
        - createdAt
        - updatedAt
...
    RepoStats:
      type: object
      properties:
        repoId:
          type: integer
        totalCommits:
          type: integer
        analyzedCommits:
          type: integer
        avgComplexity:
          type: number
        maxComplexity:
          type: number
        avgCoupling:
          type: number
        totalFiles:
          type: integer
        spikeCount:
          type: integer
        mostChangedFile:
          type: ["string", "null"]
        dateRange:
          type: object
          properties:
            start:
              type: ["string", "null"]
            end:
              type: ["string", "null"]
      required:
        - repoId
        - totalCommits
        - analyzedCommits
        - avgComplexity
        - maxComplexity
        - avgCoupling
        - totalFiles
        - spikeCount
    CommitSummary:
      type: object
      properties:
        sha:
          type: string
        shortSha:
          type: string
        message:
          type: string
        author:
          type: string
        timestamp:
          type: string
          format: date-time
        complexity:
          type: ["number", "null"]
        coupling:
          type: ["number", "null"]
        churnRisk:
          type: ["number", "null"]
        filesChanged:
          type: ["integer", "null"]
        linesAdded:
          type: ["integer", "null"]
        linesRemoved:
          type: ["integer", "null"]
        isSpikeCommit:
          type: boolean
      required:
        - sha
        - shortSha
        - message
        - author
        - timestamp
        - isSpikeCommit
    CommitPage:
      type: object
      properties:
        commits:
          type: array
          items:
            $ref: "#/components/schemas/CommitSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - commits
        - total
        - limit
        - offset
    CommitDetail:
      type: object
      properties:
        sha:
          type: string
        shortSha:
          type: string
        message:
          type: string
        author:
          type: string
        timestamp:
          type: string
          format: date-time
        complexity:
          type: ["number", "null"]
        coupling:
          type: ["number", "null"]
        churnRisk:
          type: ["number", "null"]
        filesChanged:
          type: ["integer", "null"]
        linesAdded:
          type: ["integer", "null"]
        linesRemoved:
          type: ["integer", "null"]
        isSpikeCommit:
          type: boolean
        complexityDelta:
          type: ["number", "null"]
        couplingDelta:
          type: ["number", "null"]
        changedFiles:
          type: array
          items:
            type: string
        topComplexFunctions:
          type: array
          items:
            $ref: "#/components/schemas/FunctionMetric"
      required:
        - sha
        - shortSha
        - message
        - author
        - timestamp
        - isSpikeCommit
        - changedFiles
        - topComplexFunctions
    FunctionMetric:
      type: object
      properties:
        name:
          type: string
        file:
          type: string
        complexity:
          type: number
        lineNumber:
          type: ["integer", "null"]
      required:
        - name
        - file
        - complexity
    TimelinePoint:
      type: object
      properties:
        sha:
          type: string
        shortSha:
          type: string
        timestamp:
          type: string
          format: date-time
        commitIndex:
          type: integer
        complexity:
          type: ["number", "null"]
        coupling:
          type: ["number", "null"]
        churnRisk:
          type: ["number", "null"]
        filesChanged:
          type: ["integer", "null"]
        isSpikeCommit:
          type: boolean
        message:
          type: string
        author:
          type: string
      required:
        - sha
        - shortSha
        - timestamp
        - commitIndex
        - isSpikeCommit
        - message
        - author
    HotspotFile:
      type: object
      properties:
        filePath:
          type: string
        churnCount:
          type: integer
        couplingScore:
          type: number
        complexityScore:
          type: number
        riskScore:
          type: number
        lastChangedSha:
          type: ["string", "null"]
        lastChangedAt:
          type: ["string", "null"]
      required:
        - filePath
        - churnCount
        - couplingScore
        - complexityScore
        - riskScore
...