# Repository_health_analyser/lib/api-zod/src/generated/types/commitSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RepoHealth API — git ingestion, code metrics, knowledge graphs, and LLM insights
 * OpenAPI spec version: 0.1.0
 */

export interface CommitSummary {
  sha: string;
  shortSha: string;
  message: string;
  author: string;
  timestamp: Date;
  /** @nullable */
  complexity?: number | null;
  /** @nullable */
  coupling?: number | null;
  /** @nullable */
  churnRisk?: number | null;
  /** @nullable */
  filesChanged?: number | null;
  /** @nullable */
  linesAdded?: number | null;
  /** @nullable */
  linesRemoved?: number | null;
  isSpikeCommit: boolean;
}