# Repository_health_analyser/lib/api-zod/src/generated/types/repoStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RepoHealth API — git ingestion, code metrics, knowledge graphs, and LLM insights
 * OpenAPI spec version: 0.1.0
 */
import type { RepoStatsDateRange } from './repoStatsDateRange';

export interface RepoStats {
  repoId: number;
  totalCommits: number;
  analyzedCommits: number;
  avgComplexity: number;
  maxComplexity: number;
  avgCoupling: number;
  totalFiles: number;
  spikeCount: number;
  /** @nullable */
  mostChangedFile?: string | null;
  dateRange?: RepoStatsDateRange;
}