# Repository_health_analyser/lib/api-zod/src/generated/types/timelinePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RepoHealth API — git ingestion, code metrics, knowledge graphs, and LLM insights
 * OpenAPI spec version: 0.1.0
 */

export interface TimelinePoint {
  sha: string;
  shortSha: string;
  timestamp: Date;
  commitIndex: number;
  /** @nullable */
  complexity?: number | null;
  /** @nullable */
  coupling?: number | null;
  /** @nullable */
  churnRisk?: number | null;
  /** @nullable */
  filesChanged?: number | null;
  isSpikeCommit: boolean;
  message: string;
  author: string;
}