# Repository_health_analyser/python_backend/routers/metrics.py
"""
routers/metrics.py — Code health metric endpoints.

Endpoints:
  GET /api/repos/{id}/stats          → aggregate repository stats
  GET /api/repos/{id}/commits        → list all commits with metrics
  GET /api/repos/{id}/commits/{sha}  → single commit with full detail
  GET /api/repos/{id}/timeline       → time-series health data (for charts)
  GET /api/repos/{id}/hotspots       → file-level hotspot risk map
  GET /api/repos/{id}/riskiest-files → top N highest-risk files

All responses are shaped to match the OpenAPI spec exactly so the
React Query hooks generated by Orval deserialise correctly.
"""

import json
from typing import List, Optional
from fastapi import APIRouter, Depends, HTTPException
from pydantic import BaseModel
from sqlalchemy.orm import Session
from sqlalchemy import func

from database import get_db
from models import Repository, Commit, FileMetric

router = APIRouter(prefix="/repos", tags=["metrics"])


# ---------------------------------------------------------------------------
# Response models
# ---------------------------------------------------------------------------

class FunctionMetric(BaseModel):
    name: str
    file: str
    complexity: float
    lineNumber: Optional[int] = None


class CommitSummaryResponse(BaseModel):
    sha: str
    shortSha: str
    message: str
    author: str
    timestamp: str
    complexity: Optional[float] = None
    coupling: Optional[float] = None
    churnRisk: Optional[float] = None
    filesChanged: Optional[int] = None
    linesAdded: Optional[int] = None
    linesRemoved: Optional[int] = None
    isSpikeCommit: bool


class CommitPageResponse(BaseModel):
    commits: List[CommitSummaryResponse]
    total: int
    limit: int
    offset: int


class CommitDetailResponse(CommitSummaryResponse):
    complexityDelta: Optional[float] = None
    couplingDelta: Optional[float] = None
    changedFiles: List[str] = []
    topComplexFunctions: List[FunctionMetric] = []


class TimelinePointResponse(BaseModel):
    sha: str
    shortSha: str
    timestamp: str
    commitIndex: int
    complexity: Optional[float] = None
    coupling: Optional[float] = None
    churnRisk: Optional[float] = None
    filesChanged: Optional[int] = None
    isSpikeCommit: bool
    message: str
    author: str


class HotspotFileResponse(BaseModel):
    filePath: str
    churnCount: int
    couplingScore: float
    complexityScore: float
    riskScore: float
    lastChangedSha: Optional[str] = None
    lastChangedAt: Optional[str] = None


class DateRange(BaseModel):
    start: Optional[str] = None
    end: Optional[str] = None


class RepoStatsResponse(BaseModel):
    repoId: int
    totalCommits: int
    analyzedCommits: int
    avgComplexity: float
    maxComplexity: float
    avgCoupling: float
    totalFiles: int
    spikeCount: int
    mostChangedFile: Optional[str] = None
    dateRange: DateRange


# ---------------------------------------------------------------------------
# Helpers
# ---------------------------------------------------------------------------

def _commit_to_summary(c: Commit) -> CommitSummaryResponse:
    return CommitSummaryResponse(
        sha=c.sha,
        shortSha=c.short_sha,
        message=(c.message or "")[:200],
        author=c.author or "",
        timestamp=c.timestamp.isoformat() + "Z",
        complexity=c.complexity,
        coupling=c.coupling,
        churnRisk=c.churn_risk,
        filesChanged=c.files_changed,
        linesAdded=c.lines_added,
        linesRemoved=c.lines_removed,
        isSpikeCommit=bool(c.is_spike_commit),
    )


def _require_repo(repo_id: int, db: Session) -> Repository:
    repo = db.query(Repository).filter(Repository.id == repo_id).first()
    if not repo:
        raise HTTPException(404, detail="Repository not found")
    return repo


# ---------------------------------------------------------------------------
# Endpoints
# ---------------------------------------------------------------------------

@router.get("/{repo_id}/stats", response_model=RepoStatsResponse)
def get_repo_stats(repo_id: int, db: Session = Depends(get_db)):
    """
    Return aggregate statistics for a repository.
    Used on the repo overview page header cards.
    """
    repo = _require_repo(repo_id, db)

    commits = db.query(Commit).filter(Commit.repo_id == repo_id).all()
    if not commits:
        return RepoStatsResponse(
            repoId=repo_id,
            totalCommits=repo.total_commits or 0,
            analyzedCommits=0,
            avgComplexity=0.0,
            maxComplexity=0.0,
            avgCoupling=0.0,
            totalFiles=0,
            spikeCount=0,
            dateRange=DateRange(),
        )

    complexities = [c.complexity for c in commits if c.complexity is not None]
    couplings = [c.coupling for c in commits if c.coupling is not None]
    spike_count = sum(1 for c in commits if c.is_spike_commit)

    # Most changed file
    file_metrics = (
        db.query(FileMetric)
        .filter(FileMetric.repo_id == repo_id)
        .order_by(FileMetric.churn_count.desc())
        .first()
    )
    most_changed = file_metrics.file_path if file_metrics else None
    total_files = db.query(FileMetric).filter(FileMetric.repo_id == repo_id).count()

    timestamps = [c.timestamp for c in commits if c.timestamp is not None]
    date_range = DateRange(
        start=min(timestamps).isoformat() + "Z" if timestamps else None,
        end=max(timestamps).isoformat() + "Z" if timestamps else None,
    )

    return RepoStatsResponse(
        repoId=repo_id,
        totalCommits=repo.total_commits or len(commits),
        analyzedCommits=len(commits),
        avgComplexity=round(sum(complexities) / len(complexities), 2) if complexities else 0.0,
        maxComplexity=round(max(complexities), 2) if complexities else 0.0,
        avgCoupling=round(sum(couplings) / len(couplings), 2) if couplings else 0.0,
        totalFiles=total_files,
        spikeCount=spike_count,
        mostChangedFile=most_changed,
        dateRange=date_range,
    )


@router.get("/{repo_id}/commits", response_model=CommitPageResponse)
def list_commits(repo_id: int, db: Session = Depends(get_db)):
    """
    Return all commits with their health metrics, in chronological order.
    Paginated to 500 commits max per request.
    """
    _require_repo(repo_id, db)

    total = db.query(Commit).filter(Commit.repo_id == repo_id).count()
    commits = (
        db.query(Commit)
        .filter(Commit.repo_id == repo_id)
        .order_by(Commit.commit_index)
        .limit(500)
        .all()
    )

    return CommitPageResponse(
        commits=[_commit_to_summary(c) for c in commits],
        total=total,
        limit=500,
        offset=0,
    )


@router.get("/{repo_id}/commits/{sha}", response_model=CommitDetailResponse)
def get_commit(repo_id: int, sha: str, db: Session = Depends(get_db)):
    """
    Return full details for a single commit including changed files and
    top complex functions.
    """
    _require_repo(repo_id, db)

    commit = (
        db.query(Commit)
        .filter(Commit.repo_id == repo_id, Commit.sha == sha)
        .first()
    )
    # Try short SHA lookup if full SHA not found
    if not commit:
        commit = (
            db.query(Commit)
            .filter(Commit.repo_id == repo_id, Commit.short_sha == sha)
            .first()
        )
    if not commit:
        raise HTTPException(404, detail="Commit not found")

    changed_files = []
    if commit.changed_files_json:
        try:
            changed_files = json.loads(commit.changed_files_json)
        except Exception:
            pass

    top_functions_raw = []
    if commit.top_functions_json:
        try:
            top_functions_raw = json.loads(commit.top_functions_json)
        except Exception:
            pass

    top_functions = [
        FunctionMetric(
            name=f.get("name", "unknown"),
            file=f.get("file", ""),
            complexity=f.get("complexity", 1.0),
            lineNumber=f.get("line_number"),
        )
        for f in top_functions_raw
    ]

    return CommitDetailResponse(
        sha=commit.sha,
        shortSha=commit.short_sha,
        message=(commit.message or "")[:500],
        author=commit.author or "",
        timestamp=commit.timestamp.isoformat() + "Z",
        complexity=commit.complexity,
        coupling=commit.coupling,
        churnRisk=commit.churn_risk,
        filesChanged=commit.files_changed,
        linesAdded=commit.lines_added,
        linesRemoved=commit.lines_removed,
        isSpikeCommit=bool(commit.is_spike_commit),
        complexityDelta=commit.complexity_delta,
        couplingDelta=commit.coupling_delta,
        changedFiles=changed_files[:50],
        topComplexFunctions=top_functions,
    )


@router.get("/{repo_id}/timeline", response_model=List[TimelinePointResponse])
def get_timeline(repo_id: int, db: Session = Depends(get_db)):
    """
    Return the full health timeline for charting.
    Returns up to 200 commits, evenly sampled if there are more.
    Used to power the Recharts line chart on the repo overview page.
    """
    _require_repo(repo_id, db)

    commits = (
        db.query(Commit)
        .filter(Commit.repo_id == repo_id)
        .order_by(Commit.commit_index)
        .all()
    )

    if not commits:
        return []

    # Downsample to 200 points if needed
    max_points = 200
    if len(commits) > max_points:
        step = len(commits) / max_points
        commits = [commits[int(i * step)] for i in range(max_points)]

    return [
        TimelinePointResponse(
            sha=c.sha,
            shortSha=c.short_sha,
            timestamp=c.timestamp.isoformat() + "Z",
            commitIndex=c.commit_index,
            complexity=c.complexity,
            coupling=c.coupling,
            churnRisk=c.churn_risk,
            filesChanged=c.files_changed,
            isSpikeCommit=bool(c.is_spike_commit),
            message=(c.message or "")[:80],
            author=c.author or "",
        )
        for c in commits
    ]


@router.get("/{repo_id}/hotspots", response_model=List[HotspotFileResponse])
def get_hotspots(repo_id: int, db: Session = Depends(get_db)):
    """
    Return all files ranked by composite risk score (churn × coupling).
    Used for the hotspot heatmap page.
    """
    _require_repo(repo_id, db)

    files = (
        db.query(FileMetric)
        .filter(FileMetric.repo_id == repo_id)
        .order_by(FileMetric.risk_score.desc())
        .limit(50)
        .all()
    )

    return [
        HotspotFileResponse(
            filePath=f.file_path,
            churnCount=f.churn_count,
            couplingScore=round(f.coupling_score, 3),
            complexityScore=round(f.complexity_score, 3),
            riskScore=round(f.risk_score, 4),
            lastChangedSha=f.last_changed_sha,
            lastChangedAt=f.last_changed_at.isoformat() + "Z" if f.last_changed_at else None,
        )
        for f in files
    ]


@router.get("/{repo_id}/riskiest-files", response_model=List[HotspotFileResponse])
def get_riskiest_files(repo_id: int, db: Session = Depends(get_db)):
    """
    Return the top 10 highest-risk files. Used on the repo overview page
    as the "danger zone" quick summary.
    """
    _require_repo(repo_id, db)

    files = (
        db.query(FileMetric)
        .filter(FileMetric.repo_id == repo_id)
        .order_by(FileMetric.risk_score.desc())
        .limit(10)
        .all()
    )

    return [
        HotspotFileResponse(
            filePath=f.file_path,
            churnCount=f.churn_count,
            couplingScore=round(f.coupling_score, 3),
            complexityScore=round(f.complexity_score, 3),
            riskScore=round(f.risk_score, 4),
            lastChangedSha=f.last_changed_sha,
            lastChangedAt=f.last_changed_at.isoformat() + "Z" if f.last_changed_at else None,
        )
        for f in files
    ]